transition predictor: follow weight dtype (DeepSpeed bf16 casts it), return fp32
DeepSpeed with bf16 enabled casts every submodule to bf16, so the fp32
source states hit a bf16 LayerNorm ("expected scalar type Float but found
BFloat16"). Cast the input to the weights' dtype and return fp32 so the
cosine loss stays in fp32.

+++ SpectralGuidedLearning/src/transition_loss.py
@@ -29,7 +29,10 @@ class TransitionPredictor(nn.Module):
         self.down = nn.Linear(hidden, d_model)
 
     def forward(self, s: torch.Tensor) -> torch.Tensor:
-        return self.down(F.gelu(self.up(self.norm(s))))
+        # DeepSpeed bf16 casts every submodule (this one included) to bf16 while the source
+        # states arrive as fp32; follow the weights' dtype and hand back fp32 for the cosine.
+        s = s.to(self.up.weight.dtype)
+        return self.down(F.gelu(self.up(self.norm(s)))).float()
 
 
 @dataclass
